+++ backend/springboot/src/main/java/com/workmaite/domain/audit/repository/AuditLogRepository.java
@@ -3,4 +3,4 @@ package com.workmaite.domain.audit.repository;
 import com.workmaite.domain.audit.entity.AuditLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditLogRepository extends JpaRepository<AuditLog, Integer> {}
+public interface AuditLogRepository extends JpaRepository<AuditLog, Long> {}
