+++ backend/springboot/src/main/java/com/workmaite/domain/logs/entity/HitlReview.java
@@ -15,44 +15,44 @@ public class HitlReview {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
+  private Long id;
 
   @Column(name = "agent_log_id")
-  private Integer agentLogId;
+  private Long agentLogId;
 
   @Column(name = "target_type", length = 30, nullable = false)
   private String targetType;
 
-    @Column(name = "ai_rationale", columnDefinition = "TEXT")
-    private String aiRationale;
+  @Column(name = "ai_rationale", columnDefinition = "TEXT")
+  private String aiRationale;
 
   @Builder.Default
   @Column(length = 20, nullable = false)
   private String status = "pending";
 
   @Column(name = "reviewer_id")
-  private Integer reviewerId;
+  private Long reviewerId;
 
-    @Column(name = "reviewed_at")
-    private LocalDateTime reviewedAt;
+  @Column(name = "reviewed_at")
+  private LocalDateTime reviewedAt;
 
   @CreationTimestamp
   @Column(name = "created_at", updatable = false)
   private LocalDateTime createdAt;
 
-    @Column(name = "agenda_id")
-    private Long agendaId;
+  @Column(name = "agenda_id")
+  private Long agendaId;
 
-    @Column(name = "report_id")
-    private Long reportId;
+  @Column(name = "report_id")
+  private Long reportId;
 
-    @Column(name = "comment", columnDefinition = "TEXT")
-    private String comment;
+  @Column(name = "comment", columnDefinition = "TEXT")
+  private String comment;
 
-    public void review(String status, Long reviewerId, String comment) {
-        this.status = status;
-        this.reviewerId = reviewerId;
-        this.comment = comment;
-        this.reviewedAt = LocalDateTime.now();
-    }
+  public void review(String status, Long reviewerId, String comment) {
+    this.status = status;
+    this.reviewerId = reviewerId;
+    this.comment = comment;
+    this.reviewedAt = LocalDateTime.now();
+  }
 }
