+++ backend/springboot/src/main/java/com/workmaite/domain/reports/repository/ReportScoreRepository.java
@@ -4,6 +4,6 @@ import com.workmaite.domain.reports.entity.ReportScore;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReportScoreRepository extends JpaRepository<ReportScore, Integer> {
-  Optional<ReportScore> findByReportId(Integer reportId);
+public interface ReportScoreRepository extends JpaRepository<ReportScore, Long> {
+  Optional<ReportScore> findByReportId(Long reportId);
 }
